+++ articles/tutorials/building_2d_games/22_the_game/snippets/game_scene_create_game_over_menu.cs
