+++ articles/tutorials/building_2d_games/22_the_game/snippets/game_scene_move_bat.cs
