+++ articles/tutorials/building_2d_games/22_the_game/snippets/slime_segment.cs
